requirements.txt
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==20.6
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
urllib3==2.5.0
wheel==0.45.1
python-telegram-bot[webhooks]==20.6
requests>=2.32.5
python-dotenv>=1.2.1
# Ensures zoneinfo database available in environments lacking OS tzdata
tzdata>=2024.1
# Optional: fallback timezone helper, useful in older deployments
pytz>=2024.3

# Developer convenience (optional)
setuptools>=60.0
wheel>=0.45.1
